requirements.txt
PyJWT
jwcrypto
requests
